Refactoring and cleaning up models
