+++ backend/app/models/__init__.py
@@ -1,8 +1,23 @@
 # type: ignore
 from app.models.asset import Asset as Asset
-from app.models.base import Base, ScanStatus, Severity
+from app.models.base import Base
 from app.models.finding import Finding as Finding
+from app.models.report import Report as Report
 from app.models.scan import Scan as Scan
+from app.models.scan_schedule import ScanSchedule as ScanSchedule
+from app.models.scan_source import ScanSource as ScanSource
 from app.models.user import User as User
+from app.models.verified_domain import VerifiedDomain as VerifiedDomain
 
-__all__ = ["Base", "ScanStatus", "Severity", "Asset", "Finding", "Scan", "User"]
+__all__ = [
+    "Asset",
+    "Base", 
+    "Finding",
+    #"Organisation",
+    "Report",
+    "Scan",
+    "ScanSchedule",
+    "ScanSource",
+    "User",
+    "VerifiedDomain",
+]
