+++ backend/app/models/base.py
@@ -19,3 +19,38 @@ class Severity(str,enum.Enum):
     MEDIUM = 'medium'
     HIGH = 'high'
     CRITICAL = 'critical'
+
+class ReportStatus(str, enum.Enum):
+    PENDING = "pending"
+    GENERATING = "generating"
+    COMPLETED = "completed"
+    FAILED = "failed"
+
+class ScanSourceStatus(str, enum.Enum):
+    PENDING = "pending"
+    RUNNING = "running"
+    COMPLETED = "completed"
+    FAILED = "failed"
+    PARTIAL = "partial"
+    SKIPPED = "skipped"
+
+class DomainVerificationStatus(str, enum.Enum):
+    PENDING = "pending"
+    VERIFIED = "verified"
+    FAILED = "failed"
+    EXPIRED = "expired"
+
+class DomainVerificationCode(str, enum.Enum):
+    VERIFIED = "verified"
+    RECORD_NOT_FOUND = "record_not_found"
+    TOKEN_MISMATCH = "token_mismatch"
+    LOOKUP_FAILED = "lookup_failed"
+
+class ScanType(str, enum.Enum):
+    PASSIVE_CTEM = "passive_ctem"
+    ACTIVE_VULNERABILITY = "active_vulnerability"
+
+class ScanScheduleFrequency(str, enum.Enum):
+    DAILY = "daily"
+    WEEKLY = "weekly"
+    MONTHLY = "monthly"
