--- backend/app/models/report_status.py
@@ -1,8 +1,0 @@
-import enum
-
-
-class ReportStatus(str, enum.Enum):
-    PENDING = "pending"
-    GENERATING = "generating"
-    COMPLETED = "completed"
-    FAILED = "failed"
