+++ backend/app/models/scan_source.py
@@ -5,16 +5,7 @@ from sqlalchemy import Column, DateTime, Enum, ForeignKey, String, Text, UniqueC
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import relationship
 
-from app.models.base import Base
-
-
-class ScanSourceStatus(str, enum.Enum):
-    PENDING = "pending"
-    RUNNING = "running"
-    COMPLETED = "completed"
-    FAILED = "failed"
-    PARTIAL = "partial"
-    SKIPPED = "skipped"
+from app.models.base import Base, ScanSourceStatus
 
 
 class ScanSource(Base):
